refactor add functionality of APIs
