+++ src/main/kotlin/us/timinc/mc/cobblemon/counter/api/KoApi.kt
@@ -2,10 +2,38 @@ package us.timinc.mc.cobblemon.counter.api
 
 import com.cobblemon.mod.common.Cobblemon
 import net.minecraft.entity.player.PlayerEntity
+import net.minecraft.text.Text
+import us.timinc.mc.cobblemon.counter.Counter.config
+import us.timinc.mc.cobblemon.counter.Counter.info
 import us.timinc.mc.cobblemon.counter.store.KoCount
 import us.timinc.mc.cobblemon.counter.store.KoStreak
 
 object KoApi {
+    fun add(player: PlayerEntity, species: String) {
+        val data = Cobblemon.playerData.get(player)
+        val koCount: KoCount = data.extraData.getOrPut(KoCount.NAME) { KoCount() } as KoCount
+        val koStreak: KoStreak = data.extraData.getOrPut(KoStreak.NAME) { KoStreak() } as KoStreak
+
+        koCount.add(species)
+        koStreak.add(species)
+
+        val newCount = koCount.get(species)
+        val newStreak = koStreak.get(species)
+
+        info(
+            "Player ${player.displayName.string} KO'd a $species streak($newStreak) count($newCount)"
+        )
+        if (config.broadcastKosToPlayer) {
+            player.sendMessage(
+                Text.translatable(
+                    "counter.ko.confirm", species, newCount, newStreak
+                )
+            )
+        }
+
+        Cobblemon.playerData.saveSingle(data)
+    }
+
     fun getTotal(player: PlayerEntity): Int {
         val playerData = Cobblemon.playerData.get(player)
         return (playerData.extraData.getOrPut(KoCount.NAME) { KoCount() } as KoCount).total()
